refactor: Move OwinFixture out of OwinWebApi2Tests
